Add basic frontend for site login
