+++ src/lib/state/auth.svelte.ts
@@ -1,0 +1,3 @@
+const authState = $state({ isAuthenticated: false });
+
+export default authState;
